FI-943 fix: use CDP only when `cdpClient` is defined
fix: mark `RequestHookWithEvents` as internal class
refactor: split `SetHeadersRequestHook` into several modules
