+++ src/context/cdpClient.ts
@@ -1,5 +1,5 @@
 import {useContext} from '../useContext';
-import {assertValueIsDefined, assertValueIsUndefined} from '../utils/asserts';
+import {assertValueIsUndefined} from '../utils/asserts';
 
 import type {CdpClient} from '../types/internal';
 
@@ -7,26 +7,20 @@ import type {CdpClient} from '../types/internal';
  * Raw versions of `getCdpClient` and `setCdpClient`.
  * @internal
  */
-const [getRawCdpClient, setRawCdpClient] = useContext<CdpClient>();
+const [getCdpClient, setRawCdpClient] = useContext<CdpClient>();
 
 /**
  * Get test run `cdpClient`.
  * @internal
  */
-export const getCdpClient = (): CdpClient => {
-  const cdpClient = getRawCdpClient();
-
-  assertValueIsDefined(cdpClient, 'cdpClient is defined');
-
-  return cdpClient;
-};
+export {getCdpClient};
 
 /**
  * Set test `cdpClient` (can only be called once).
  * @internal
  */
 export const setCdpClient: typeof setRawCdpClient = (cdpClient) => {
-  const currentCdpClient = getRawCdpClient();
+  const currentCdpClient = getCdpClient();
 
   assertValueIsUndefined(currentCdpClient, 'currentCdpClient is not defined', {cdpClient});
