+++ src/utils/requestHooks/setHeaders/index.ts
@@ -1,0 +1,2 @@
+/** @internal */
+export {SetHeadersRequestHook} from './SetHeadersRequestHook';
